tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/animation_shorthand_after_longhand.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - animation shorthand after animation longhand</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#the-animation-shorthand-property-">
   <meta name="flags" content="dom">
   <meta name="assert" content="Omitted values in the animation shorthand still
   override animation-* longhand properties with lower weight in the cascade">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <style type="text/css">
      @keyframes move {
         from {
            left: 0px;
         }
         to {
            left: 100px;
         }
      }
      div {
         width: 100px;
         height: 100px;
         background: blue;
         position: relative;
      }
      .test {
         animation-fill-mode: forwards;
         animation: step-start 700ms 0.1s move;
      }
   </style>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var div = document.querySelector('div');
         var log = document.querySelectorAll('p')[1].firstChild;
         var style = getComputedStyle(div, null);
         var passed = false;
         setTimeout(function() {
            passed = (style.left == '100px');
         }, 400);
         setTimeout(function() {
            passed = (passed && style.left == 'auto');
            test(function(){assert_true(passed);}, testName); done();
            log.data = (passed ? 'PASS if the animation occurred.' : 'FAIL');
         }, 1000);
         div.className = 'test';
      }
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>The blue square below should move one step rightwards, then back left
   again.</p>
   <p>Testing...</p>
   <div></div>
</body>
</html>
